added formulas for shape types to skeleton

+++ Assign1/Shapes/Shape.java
@@ -3,12 +3,12 @@ package Shapes;
 import java.util.Comparator;
 
 public abstract class Shape implements Comparable<Shape>, Comparator<Shape> {
-    public int height;
-    public int volume;
-    public int baseArea;
+    public double height;
+    public double volume;
+    public double baseArea;
 
     @Override
-    public int compareTo(Shape s) {
+    public double compareTo(Shape s) {
         return compare(this, s);
     }
